extract db operations into their own files + eliminate inheritance
