+++ core/src/main/scala/org/broadinstitute/dsde/rawls/monitor/migration/Shared.scala
@@ -1,0 +1,33 @@
+package org.broadinstitute.dsde.rawls.monitor.migration
+
+import cats.implicits.{catsSyntaxOptionId, toTraverseOps}
+import org.broadinstitute.dsde.rawls.RawlsException
+
+object Shared {
+  sealed trait Outcome
+
+  object Outcome {
+    case object Success extends Outcome
+
+    final case class Failure(message: String) extends Outcome
+
+    final def fromFields(outcome: Option[String], message: Option[String]): Either[String, Option[Outcome]] = {
+      type EitherStringT[T] = Either[String, T]
+      outcome.traverse[EitherStringT, Outcome] {
+        case "Success" => Right(Success)
+        case "Failure" => Right(Failure(message.getOrElse("")))
+        case other => Left(s"""Failed to read outcome: unknown value -- "$other"""")
+      }
+    }
+
+    final def toTuple(outcome: Outcome): (Option[String], Option[String]) = outcome match {
+      case Success => ("Success".some, None)
+      case Failure(msg) => ("Failure".some, msg.some)
+    }
+  }
+
+  def unsafeFromEither[A, B](f: A => Either[String, B], a: A): B = f(a) match {
+    case Right(r) => r
+    case Left(msg) => throw new RawlsException(msg)
+  }
+}
